Added Blip Wrapper and Prop Class for Objects
The purpose of the Blip class is to make handling Blips easier. It is
still missing many Blip Natives and should be extended in the future.

The purpose of the Prop class is to enable the management of (static)
Objects in the game. Objects are Entities, but not Peds or Vehicles.
Since Entity is abstract we have no means of handling them at the
moment. There is a (most likely incomplete) list of object hashes (
ecb2.biz/releases/GTAV/lists/ )which I could be included as an enum
(PropHash - i.e. in the same way we have the PedHash and VehicleHash
enum).