Added ability to set bone offsets in realtime. Fixed some typos in the EntityBone class.
Added two new properties to the EntityBone class. "Pose" and
"PoseMatrix". Adjusting them allows setting the relative position of a
bone in realtime. This is based off how R* does it (such as with the
ruiner2 missiles). Also added a few missing Read and Write functions to
the MemoryAccess class.