Return correct weapon hash instead of ammo type hash in WeaponHash
EXPLODE_PROJECTILES actually explodes all the projectiles associated
with the ammo type hash from the weapon hash (e.g. compact grenade
launcher and the grenade launcher that is in the base game).

The new weapon hash value is used when dealing damage with explosion
to entities. If the value is not matched to any weapon hash,
the explosion will not deal any damage to any entities.