Fix Ped.CanBeKnockedOffBike
SET_PED_CAN_BE_KNOCKED_OFF_VEHICLE uses an enum, not a bool, and the default values are '0' for default (can be knocked off bike) or '1' for never (cannot be knocked off bike). As a result, inverting the bool passed to Ped.CanBeKnockedOffBike will provide the user the result that makes logical sense (namely that setting it to true will allow peds to be knocked off, and false will not allow peds to be knocked off).