Bulk add kagikn as a copyright author
Update SHVDN license URLs

Since I am the project lead now, I might as well be entitled to be listed
as an author explicitly so you can know who's the current lead, correct?
When the organization grows enough, the author could be renamed as like
the SHVDN project or collective, though.