feat(api-v3): OutputArgument.Init (#1268)
* feat(api-v3): OutputArgument.Init

This commit introduces ways to initialize OutputArguments with primitive
values and bittable structs in v3 API, making it possible to use native
functions that requires pointer addresses without having to use members
of the Marshal class or use unsafe blocks in C#.

Although you can init OutputArgument with a primitive value, string,
or class that implements IHandleable in in v2 API, you could init one
only with a string and classes that inherits INativeValue in v3 API
before.

This commit introduces Fody and InlineIL.Fody as dev-dependencies (not
required to place as separate files). This is done because we could not
replicate Unsafe.AsPointer of System.Runtime.CompilerServices using
ILGenerator and DynamicMethod and Func types. There are no ways to use
generic pointer types as constraint for Func generic types even with
the unmanaged constraint.