feat(api-v3): dedicated FeedItem class, more feed post methods, and HudColor enum (#1284)
I feel like we could add a readonly struct that represents a pair of a
texture dictionary and texture name so we can handle txd stuff easily,
but I don't have any ideas of good names for that struct type.
TexturePath? TxdAndTextureNamePair? Those sounded a bit awkward to me,
but maybe wrong assumption.