feat(api-v3): new World.ShootSingleBullet methods
Way more documented than existing World.ShootBullet method.
I chose not to declare the weapon parameter as a mandatory one despite
the equivalent one being optional in the official header for ysc scripts
(with the default value WEAPONTYPE_ASSAULTRIFLE).

I thought we could introduce the first builder class, but I decided not
to do for shoot single bullet methods as we have named arguments and
optional parameters that can omit explicit value specification in C#,
which is different from Java and Rust (for methods with a lot of
mandatory parameters, we could use Type-safe builders in conjuction to
generics with phantom data).