fix(api-v3): avoid potential ArrayTypeMismatchException caused by what Get*PickupObjects methods return
Now GetAllPickupObjects, GetNearbyPickupObjects, and
GetClosestPickupObject return Prop[] as in v3.6.0 and earlier versions.
Arrays in .NET is co-variant, so the change had potential factors to be
breaking changes. Assign an int to some element of an array of string
via an object[] variable, and you fucked up.

Partially reverts dbbf7a6 (which is pushed for #1277).