feat!(api-v3): rename VehicleWheel properties that aren't exposed in stable versions and CWheel vars in NativeMemory to match internal names of the game
Well, a leak incident on some debug build or tool happened in 2023.
I missed the leak, but some prominent person passed the archive file
for being in the community long enough (and thus making me trustworthy
for him).

And now I correct names associated with the CWheel struct/class as far
as the compatibilities for scripts build against v3.6.0 or a earlier ver
won't break.