feat!(api-v3): move 3 static water methods to new Water static class and rename to the same internal names in exe
A leak of roadpathsgenerator_win64_beta probes that there are static
Water class in GTA5, which contains GetWaterLevel, GetWaterLevelNoWaves,
and TestLineAgainstWater, and I thought it would be great to move
relevant water methods in our codebase into a new Water static class
so the World class won't get NAIVELY bloated.

Also renames the 3 methods to the same internal names defined in the
exe.