feat(core): add the ability display outputs with multiple lines to console
Resolves #1301
This commit adds the ability to display output strings with multiple
lines to the console, making output strings will not always get
overlapped even if there is no room for displaying in single line.

Uses binary search to avoid taking too long time to calculate where to
split strings.

Do note that this commit itself does not make it possible to display
long strings that needs to be pushed with more than 4 substrings where
the length is up to 99 excluding the null bytes. This is probably
because END_TEXT_COMMAND_DISPLAY_TEXT cannot display strings with more
than 4 variable substrings. Maybe ScriptTextLinesFourSubstringsThreeNumbers
in gameconfig.xml is relevant a bit?