feat(api-v3): rename 2 enums to avoid name collision and add 2 new wrapper classes for peds (#1430)
This commit is about restructuring since `CPedConfigFlags` and
`CPedResetFlags` are the things that exist along with `ePedConfigFlags`
and `ePedResetFlags` enums.

`PedConfigFlags` and `PedResetFlags` are renamed to
`PedConfigFlagToggles` and `PedResetFlagToggles` respectively now to
free the identifiers and the new wrapper classes `PedConfigFlags` and
`PedResetFlags` take it now. `PedConfigFlagToggles` and
`PedResetFlagToggles` are for `ePedConfigFlags` and `ePedResetFlags`
respectively, and the new `PedConfigFlags` and `PedResetFlags` are for
`CPedConfigFlags` and `CPedResetFlags`.

This commit also gives some of the idea how `CPedConfigFlags` and
`CPedResetFlags` would align members just in case Rockstar removes some
of them.

Git doesn't seem to detect the changes of our `PedConfigFlags` and
`PedResetFlags` too well from the histories, and the diffs show as
massive rewrites...
You can see the *actual* previous commit of
`PedConfigFlagToggles` in 51f89a942a234abe8b1aaea387fae515a0e2baac,
and you can see that of `PedResetFlagToggles` in 46d7b87f8485c028fdf33568c8cebb3fa13d8e77.