feat(api-v3): `Ped.SetIKTarget` (#1441)
## Summary
Adds `Ped.SetIKTarget` with typed IK part enum and sufficient info to
figure out how the IK target flag parameter works.

fingaweg not having reviewed PRs in the GTAV native DB increased
priority to publish this PR a lot, as well as the DB not having
sufficient info to completely understand what the native expects as the
params. While he merged 2 PRs in RDR3 one in this 1 year. We could build
docs in SHVDN for some methods or properties that use natives that
aren't fully documented in that native DB.