fix(core): don't crash game for dereferencing null console domain
This commit fixes console commands crashing the game (including when the
instant reload key is pressed), as well as the keyboard message method
sometimes reading stale values of the script and console domain
variables.
This commit also adds additional error logs when the runtime happens to
fail to call console commands for the console data missing from the
script domain instance.