feat(api-v3): `Game.FileVersion` as an alternative to `Game.Version` (#1501)
AB is deprecating SHV's `getGameVersion` and `Game.Version`, so I
thought new game version API should be introduced. I wish AB hadn't
thought of this if `getGameVersion` is widely used and the new API can't
be isolated in a new DLL, but the things are happening.

Adds `GameVersionNotSupportedException.MinimumSupportedGameVersion`
and mark `GameVersionNotSupportedException.MinimumSupportedGameVersion`
as obsolete as Script Hook V is deprecating `getGameVersion`.