ci: calculate version string for assemblies faster
The build workflow now doesn't use regex against `GITHUB_REF` unless
the version input is empty. The stable-release workflow now specifies
a version as the version input of the build workflow. The stable-release
one also runs the job that create a release draft on Ubuntu rather than
Windows because it doesn't have to run on Windows.