fix(api-v3): `Game.RadioStation` not stopping the radio even if the radio is using
Note that the property does nothing when a `RadioStation` value that is
out of range of `RadioStation` is passed in the v3 API, while it behave
like as if `RadioStation.RadioOff` is passed in the v2 API.
If the game can have radio stations that has the null radio hash,
which is zero, the property in the v3 API would retune the radio to
one, but I believe that won't happen in reality.

Fixes #1516.