feat(core): change log level of deprecated script warning to `Info`
The warning messages just scared a lot of users too much but it looks
like only few of them reported to developers who created scripts that
use the v2 API. Not a few active script devs cared the messages, though.
Changing the log level of the messages to `Info` would be justified for
that reason.