feat(api-v3): add custom `ToString()` impls to 4 game clock date time and duration structs (#1534)
Overrides `ToString()` on the 4 structs in `GTA.Chrono`, namely
`GameClockDate`, `GameClockTime`, `GameClockDateTime`, and
`GameClockDuration`.

None of the overridden `ToString()`s uses the `DateTimeFormatInfo` of 
the `CultureInfo` on the thread where the current script is running.

String Formats:
- `GameClockDate`: `YYYY-MM-DD` for dates where the year is between 0
and 9999, and `-YYYY-MM-DD` for dates where the year is negative and
more than -10000. If the year is not between -9999 and 9999, then 
`Y-MM-DD` where the year part has no padding zero characters and has
the negative sign if negative. Loosely follows the extended date
format of ISO 8601, but strictly complies with it for dates where
the years are between -9999 and 9999.
- `GameClockTime`: Always `HH:MM:SS`, where each part can have one
leading zero character. Strictly complies with the extended time format
of ISO 8601.
- `GameClockDateTime`: Combined with `GameClockDate.ToString()` and
`GameClockTime.ToString()` but with a space separator, like
`2013-9-16 20:00:00`.
- `GameClockDuration`: `HH:MM:SS` for durations that has 0 whole days,
and `D:HH:MM:SS` where the date part has no padding zero characters for
ones that have 1 or more whole days.

Closes #1533.