fix(api-v3): `Vehicle.ExhaustPopsEnabled` doing nothing in almost all cases
`Vehicle.ExhaustPopsEnabled` passed what was passed as the property
value as a vehicle handle (1st argument) and the state value (2nd
argument) as false. the state value was always passed as false because
SHV always initializes the argument values to zero when calling native
functions (which is actually different from `.#sc` scripts call them).
As a result, the property did nothing in almost all cases, or it set
the vehicle with the handle being `1` was set to false only if the true
was passed as the property value AND there was the vehicle with
the handle being `1`.

The property never did anything if the property value was false,
because `GTA.Native.Function.Call` converts `false` to the `int` value
`0` and script GUIDs (what we use as entity handles) can never be zero
as a valid one.

Fixes #1550.