feat(core): fallback to last deprecated API for raw scripts without API notations (#1551)
SHVDN now tries to compile raw scripts with the last deprecated API
version if API version is not notated by the file names and they can't
compile with the first non-deprecated API version. However, this is 
really for fool proof, and script authors or users should annotate API
versions for raw scripts, so they will work as the authors intended.