build(core): use b617 version of SHV SDK again
We didn't use any features added in the b3337 version of SHV at all.
Maybe we would use `scriptsAreLaunchedUsingReloading`, which is added
in SHV b3351 and @Tanjitsu suggested in #1549, but with dynamic checks
so users can use SHV where the latest supported game version is b3337
or older (probably). Custom start up and clean up callbacks that can
take more than 2 seconds, which is different from `DllMain`, would be
definitely worth bumping the minimal SHV version, but I still doubt if
`scriptsAreLaunchedUsingReloading` is worth it.

Resolves #1555 for eliminating the contradiction of how the SDK should
be installed.