feat!(api-v3): add `Wanted` and `DispatchData` classes (#1560)
## Summary
This commit adds `Wanted` and `DispatchData` classes so `Player` class won't
get too many responsibilities so fast. We definitely know there are
`CWanted` and `CDispatchData` even, so we can be confident with how to
separate such responsibilities thanks to the source code leak of the
game.

Also removes properties and methods that are not present in v3.6.0 and
mark ones that are present in v3.6.0 as obsolete with info about
alternative ones.

We could have deduce the presence of `CWanted` in the source code of the
GTA V without the leak from how that class is present in 3D Era games,
but we couldn't have 100% sure what native functions don't access to the
class at all in that case.