ci: use local subset of SHV SDK (#1584)
It looks like we might as well use a local subset of SDK files
downloaded from `https://ntscorp.ru` because that website is too
unreliable as a SDK source. Also remove our regular workflow that tests
if HTTP status code of the SDK URL is 404, since we don't rely on any
online source for the SDK. This commit is proposed to fix the build issue
that is happened in #1582.

`https://ntscorp.ru` is not letting clients (users) download the SHV SDK
because of its SSL certificate issues, roughly since Jury 12th, 2025
(UTC). The website lets clients the latest built binary as of the commit
is made, but not for the SDK.

This commit does not change the `.gitignore` in the sdk directory, so
developers and contributers can change some content of the SDK in it
without a lot of chore (this commit adds the SDK content with the force
option).