ci: use latest nightly version tag for CI builds for PR checks
Uses the latest nightly version tag for CI builds even when a custom
version is not supplied with a workflow call or the passed ref is not
a version tag, such as builds for PR commits. Previously, the latest
stable version tag is used to determine versions for built binaries.