feat(api-v3): `IsStuckTimerUp` and `ResetVehicleStuckTimer` to `Vehicle` (#1594)
The 2 methods may be moved to the new class `VehicleDamage` later, which
represents `CVehicleDamage` in the game codebase.

I apologize for the long detailed docs of `Vehicle.IsStuckTimerUp`,
`VehicleStuckType.HungUp`, and `VehicleStuckType.Jammed`, but I know the
internal game function `CVehicleDamage::ProcessStuckCheck(float
fTimeStep)` is that complicated to process stuck timers.