fix(core): make player ped address methods work without relevant func pointers
This commit makes `GetPlayerPedAddress` and `GetLocalPlayerPedAddress`
not crash the game when the function pointers they want are not
available. They use native functions in that case.

Although this fixes some issues caused in the Enhanced version, this
commit should also do the job for game versions for Legacy one where
the function pointers cannot be found with the memory pattern we
currently use.