ci: don't let artifact missing when immutable releases is enabled
This is works along with 78f8eaf777ee887619269303e74b6b123d2cbd22.
Without this change, each release won't have a corresponding artifacts
that have built binaries when immutable releases is enabled as it
disallows assets and tags from being modified once a release is
published and therefore artifacts have to be attached when it's a draft
one.